moved to the middle, reused 'get_weights_permute' attribute
